requirements.txt
pillow
gtts
ffmpeg-python
